Add actions.py and card numbers
Moved the non-rendering endpoints to a new file to clean up views.py. Added a card total to the home page for each deck.
