+++ website/templates/heat_map.html
@@ -1,4 +1,4 @@
-<!--home.html-->
+<!--heat_map.html-->
 <!--This line makes use of the already existing base template. We can now override
 parts of it.-->
 {%extends "base.html"%}
@@ -38,10 +38,6 @@ parts of it.-->
       <div class="col" id="December">
       </div>
   </div>
-
-
-
-
 
 {% endblock %}
