+++ website/views.py
@@ -68,7 +68,6 @@ def study(deck_id):
 
         return jsonify({'success': True}), 200
 
-
     deck = Deck.query.get_or_404(deck_id)
     flashcards = [flashcard.to_dict() for flashcard in deck.flashcards]
 
@@ -123,42 +122,6 @@ def edit_flashcard(flashcard_id):
         return jsonify({'success': True, 'flashcard': {'id': flashcard.id, 'front': flashcard.front, 'back': flashcard.back}})
 
     return render_template('edit_card.html', flashcard=flashcard)
-
-@views.route('/delete-flashcard', methods=['POST'])
-@login_required
-def delete_flashcard():
-    #This converts the json data to a python dictionary
-    flashcard_data = json.loads(request.data)
-    #This gets the flashcard id from the dictionary. The key is 'flashcardId' and the value is the acutal flashcard id
-    flashcard_id = flashcard_data['flashcardId']
-    #This gets the flashcard object using the flashcard id
-    flashcard = Flashcard.query.get(flashcard_id)
-
-    #This checks if the current user is the user who created the flashcard
-    if flashcard and flashcard.user_id == current_user.id:
-            db.session.delete(flashcard)
-            db.session.commit()
-            #This returns a json object with the deleted key set to true
-            return jsonify({ 'deleted': True })
-    
-    return jsonify({ 'deleted': False })
-
-# This works the same as the delete-flashcard function but also deletes the flashcards in the deck using the deck_id
-@views.route('/delete-deck', methods=['POST'])
-@login_required
-def delete_deck():
-    deck_data = json.loads(request.data)
-    deck_id = deck_data['deckId']
-    deck = Deck.query.get(deck_id)
-
-    if deck and deck.user_id == current_user.id:
-        # Delete all flashcards associated with the deck
-        Flashcard.query.filter_by(deck_id=deck_id).delete()
-        db.session.delete(deck)
-        db.session.commit()
-        return jsonify({'success': True})
-
-    return jsonify({'success': False}), 400
 
 # This is the backend for Heat Map. It reads the StudyFrequency table based on the user and passes the date and freq to the front end.
 @views.route('/heatmap/<int:user_id>', methods=['GET', 'POST'])
